requirements.txt
psycopg2-binary==2.9.6
beautifulsoup4==4.11.1
requests==2.28.2
